+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  - Optimize the pending block construction task
  - Simplify config files. (MIGRATION NOTE!!)
  - Add /magic list command (WIP, currently lists lost
+ - Add phase spell
 
 ## 2.8.8